proper calculation of status with large numbers

+++ corehq/ex-submodules/soil/progress.py
@@ -53,10 +53,7 @@ def get_task_progress(task):
         else:
             current = info.get('current')
             total = info.get('total')
-            percent = int(
-                current * 100. / total if total and current is not None
-                else 0
-            )
+            percent = current * 100 / total if total and current is not None else 0
     return TaskProgress(
         current=current,
         total=total,
